requirements.txt
beautifulsoup4==4.11.1
python-dotenv==1.0.1
requests==2.31.0